+++ laugh_vue/laugh/src/views/SingleLayout/Login.vue
@@ -1,0 +1,8 @@
+<template>
+  ログイン
+</template>
+<script setup>
+import { computed } from 'vue'
+import { useStore } from 'vuex'
+
+</script>
